requirements.txt
aiogram==3.15.0
sqlalchemy==2.0.27
python-dotenv==1.0.0
pandas==2.2.0
openpyxl==3.1.2
alembic==1.13.1
aiosqlite==0.19.0